requirements.txt
# Warning - for prod remove the testing and tests
Django>=6.0
djangorestframework
# Documentation 
drf-spectacular
# Production WSGI Server
gunicorn            
psycopg2-binary
django-environ
networkx
# Testing
pytest
pytest-django
# For generating test data
factory-boy
Pillow